+++ libevmasm/Assembly.cpp
@@ -353,16 +353,28 @@ void Assembly::injectStart(AssemblyItem const& _i)
 
 Assembly& Assembly::optimise(bool _enable, bool _isCreation, size_t _runs)
 {
-	optimiseInternal(_enable, _isCreation, _runs);
+	optimiseInternal(_enable, _isCreation, _runs, set<size_t>{});
 	return *this;
 }
 
-map<u256, u256> Assembly::optimiseInternal(bool _enable, bool _isCreation, size_t _runs)
+map<u256, u256> Assembly::optimiseInternal(bool _enable, bool _isCreation, size_t _runs, set<size_t> const& _tagsPushedByParent)
 {
-	for (size_t subId = 0; subId < m_subs.size(); ++subId)
+	if (!m_subs.empty())
 	{
-		map<u256, u256> subTagReplacements = m_subs[subId]->optimiseInternal(_enable, false, _runs);
-		BlockDeduplicator::applyTagReplacement(m_items, subTagReplacements, subId);
+		map<size_t, set<size_t>> pushedSubTags;
+		for (AssemblyItem const& item: m_items)
+			if (item.type() == PushTag)
+			{
+				auto subTag = item.splitForeignPushTag();
+				if (subTag.first >= m_subs.size())
+					continue;
+				pushedSubTags[subTag.first].insert(subTag.second);
+			}
+		for (size_t subId = 0; subId < m_subs.size(); ++subId)
+		{
+			map<u256, u256> subTagReplacements = m_subs[subId]->optimiseInternal(_enable, false, _runs, pushedSubTags[subId]);
+			BlockDeduplicator::applyTagReplacement(m_items, subTagReplacements, subId);
+		}
 	}
 
 	map<u256, u256> tagReplacements;
@@ -385,19 +397,19 @@ map<u256, u256> Assembly::optimiseInternal(bool _enable, bool _isCreation, size_
 			count++;
 		}
 
+		// Control flow graph that resets knowledge at path joins.
+		ControlFlowGraph cfg(m_items, _tagsPushedByParent, false);
+		AssemblyItems optimisedItems;
+		for (BasicBlock const& block: cfg.optimisedBlocks())
 		{
-			// Control flow graph optimization has been here before but is disabled because it
-			// assumes we only jump to tags that are pushed. This is not the case anymore with
-			// function types that can be stored in storage.
-			AssemblyItems optimisedItems;
-
-			auto iter = m_items.begin();
-			while (iter != m_items.end())
+			auto iter = m_items.begin() + block.begin;
+			auto const end = m_items.begin() + block.end;
+			while (iter < end)
 			{
 				KnownState emptyState;
 				CommonSubexpressionEliminator eliminator(emptyState);
 				auto orig = iter;
-				iter = eliminator.feedItems(iter, m_items.end());
+				iter = eliminator.feedItems(iter, end);
 				bool shouldReplace = false;
 				AssemblyItems optimisedChunk;
 				try
@@ -424,11 +436,11 @@ map<u256, u256> Assembly::optimiseInternal(bool _enable, bool _isCreation, size_
 				else
 					copy(orig, iter, back_inserter(optimisedItems));
 			}
-			if (optimisedItems.size() < m_items.size())
-			{
-				m_items = move(optimisedItems);
-				count++;
-			}
+		}
+		if (optimisedItems.size() < m_items.size())
+		{
+			m_items = move(optimisedItems);
+			count++;
 		}
 	}
 
@@ -550,7 +562,7 @@ LinkerObject const& Assembly::assemble() const
 			break;
 		case Tag:
 			assertThrow(i.data() != 0, AssemblyException, "");
-			assertThrow(i.splitForeignPushTag().first == size_t(-1), AssemblyException, "Foreign tag.");
+			assertThrow(!i.isForeignTag(), AssemblyException, "Foreign tag.");
 			assertThrow(ret.bytecode.size() < 0xffffffffL, AssemblyException, "Tag too large.");
 			m_tagPositionsInBytecode[size_t(i.data())] = ret.bytecode.size();
 			ret.bytecode.push_back((byte)Instruction::JUMPDEST);