+++ libevmasm/AssemblyItem.cpp
@@ -50,6 +50,12 @@ void AssemblyItem::setPushTagSubIdAndTag(size_t _subId, size_t _tag)
 	setData(_tag + (u256(_subId + 1) << 64));
 }
 
+bool AssemblyItem::isForeignTag() const
+{
+	assertThrow(m_type == PushTag || m_type == Tag, Exception, "");
+	return data() >= (u256(1) << 64);
+}
+
 unsigned AssemblyItem::bytesRequired(unsigned _addressLength) const
 {
 	switch (m_type)